requirements.txt

evalutils==0.2.3

scikit-learn==0.20.2
scipy==1.2.1
pynrrd==0.4.2
